+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 This file contains highlights of what changes on each version of the Dart Test
 Runner package.
 
+#### Version 0.2.4 ####
+
+- If no browser tests are detected `content_shell` won't be needed.
+- Added a `--skip-browser-tests` that will skip all browser tests and won't
+  require `content_shell`.
+
 #### Version 0.2.3 ####
 
 - Tweaked command line output. Now show failed test suite details by default.